Fix -Wformat= warnings when passing enum to integer specifiers https://bugs.webkit.org/show_bug.cgi?id=214441
Reviewed by Darin Adler.

Covered by existing tests.

* platform/mediastream/MediaConstraints.cpp:
(WebCore::MediaConstraint::log const):
(WebCore::BooleanConstraint::logAsBoolean const):
(WebCore::DoubleConstraint::logAsDouble const):
(WebCore::IntConstraint::logAsInt const):
* platform/mediastream/MediaConstraints.h: Made uint8_t the underlying
type for MediaConstraints::DataType.
* platform/mediastream/RealtimeMediaSource.cpp:
(WebCore::RealtimeMediaSource::selectSettings):
* platform/mediastream/RealtimeMediaSourceSupportedConstraints.h: Made
uint8_t the underlying type for MediaConstraintType.


Canonical link: https://commits.webkit.org/227282@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264534 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/mediastream/MediaConstraints.cpp
@@ -419,23 +419,23 @@ void MediaConstraint::log() const
         break;
     case DataType::None:
     case DataType::String:
-        WTFLogAlways("MediaConstraint %d of type %d", constraintType(), dataType());
+        WTFLogAlways("MediaConstraint %d of type %d", static_cast<int>(constraintType()), static_cast<int>(dataType()));
     }
 }
 
 void BooleanConstraint::logAsBoolean() const
 {
-    WTFLogAlways("BooleanConstraint %d, exact %d, ideal %d", constraintType(), m_exact ? *m_exact : -1, m_ideal ? *m_ideal : -1);
+    WTFLogAlways("BooleanConstraint %d, exact %d, ideal %d", static_cast<int>(constraintType()), m_exact ? *m_exact : -1, m_ideal ? *m_ideal : -1);
 }
 
 void DoubleConstraint::logAsDouble() const
 {
-    WTFLogAlways("DoubleConstraint %d, min %f, max %f, exact %f, ideal %f", constraintType(), m_min ? *m_min : -1, m_max ? *m_max : -1, m_exact ? *m_exact : -1, m_ideal ? *m_ideal : -1);
+    WTFLogAlways("DoubleConstraint %d, min %f, max %f, exact %f, ideal %f", static_cast<int>(constraintType()), m_min ? *m_min : -1, m_max ? *m_max : -1, m_exact ? *m_exact : -1, m_ideal ? *m_ideal : -1);
 }
 
 void IntConstraint::logAsInt() const
 {
-    WTFLogAlways("IntConstraint %d, min %d, max %d, exact %d, ideal %d", constraintType(), m_min ? *m_min : -1, m_max ? *m_max : -1, m_exact ? *m_exact : -1, m_ideal ? *m_ideal : -1);
+    WTFLogAlways("IntConstraint %d, min %d, max %d, exact %d, ideal %d", static_cast<int>(constraintType()), m_min ? *m_min : -1, m_max ? *m_max : -1, m_exact ? *m_exact : -1, m_ideal ? *m_ideal : -1);
 }
